+++ tests/integration/RequestBuilder/ItemPropertiesGetRequestBuilderTest.php
@@ -3,8 +3,10 @@
 namespace Lmc\Matej\IntegrationTests\RequestBuilder;
 
 use Lmc\Matej\IntegrationTests\IntegrationTestCase;
+use Lmc\Matej\Model\Response\ItemPropertiesListResponse;
 
 /**
+ * @covers \Lmc\Matej\Model\Response\ItemPropertiesListResponse
  * @covers \Lmc\Matej\RequestBuilder\ItemPropertiesGetRequestBuilder
  */
 class ItemPropertiesGetRequestBuilderTest extends IntegrationTestCase
@@ -18,5 +20,10 @@ class ItemPropertiesGetRequestBuilderTest extends IntegrationTestCase
             ->send();
 
         $this->assertResponseCommandStatuses($response, 'OK');
+        $this->assertInstanceOf(ItemPropertiesListResponse::class, $response);
+        $this->assertTrue($response->isSuccessful());
+        $this->assertSame('', $response->getMessage());
+        $this->assertSame('OK', $response->getStatus());
+        $this->assertInternalType('array', $response->getData());
     }
 }