+++ tests/integration/RequestBuilder/SortingRequestTest.php
@@ -6,8 +6,11 @@ use Lmc\Matej\IntegrationTests\IntegrationTestCase;
 use Lmc\Matej\Model\Command\Interaction;
 use Lmc\Matej\Model\Command\Sorting;
 use Lmc\Matej\Model\Command\UserMerge;
+use Lmc\Matej\Model\CommandResponse;
+use Lmc\Matej\Model\Response\SortingResponse;
 
 /**
+ * @covers \Lmc\Matej\Model\Response\SortingResponse
  * @covers \Lmc\Matej\RequestBuilder\SortingRequestBuilder
  */
 class SortingRequestTest extends IntegrationTestCase
@@ -20,7 +23,9 @@ class SortingRequestTest extends IntegrationTestCase
             ->sorting(Sorting::create('user-a', ['itemA', 'itemB', 'itemC']))
             ->send();
 
+        $this->assertInstanceOf(SortingResponse::class, $response);
         $this->assertResponseCommandStatuses($response, 'SKIPPED', 'SKIPPED', 'OK');
+        $this->assertShorthandResponse($response, 'SKIPPED', 'SKIPPED', 'OK');
     }
 
     /** @test */
@@ -33,6 +38,18 @@ class SortingRequestTest extends IntegrationTestCase
             ->setInteraction(Interaction::bookmark('user-a', 'item-a'))
             ->send();
 
+        $this->assertInstanceOf(SortingResponse::class, $response);
         $this->assertResponseCommandStatuses($response, 'OK', 'OK', 'OK');
+        $this->assertShorthandResponse($response, 'OK', 'OK', 'OK');
+    }
+
+    private function assertShorthandResponse(SortingResponse $response, $interactionStatus, $userMergeStatus, $sortingStatus): void
+    {
+        $this->assertInstanceOf(CommandResponse::class, $response->getInteraction());
+        $this->assertInstanceOf(CommandResponse::class, $response->getUserMerge());
+        $this->assertInstanceOf(CommandResponse::class, $response->getSorting());
+        $this->assertSame($interactionStatus, $response->getInteraction()->getStatus());
+        $this->assertSame($userMergeStatus, $response->getUserMerge()->getStatus());
+        $this->assertSame($sortingStatus, $response->getSorting()->getStatus());
     }
 }